Split classes to packages
