+++ src/com/javaguru/shoppinglist/ShoppingListApplication.java
@@ -1,12 +1,18 @@
 package com.javaguru.shoppinglist;
 
-import com.javaguru.shoppinglist.Validator.*;
+import com.javaguru.shoppinglist.database.InMemoryDatabase;
+import com.javaguru.shoppinglist.service.DefaultProductService;
+import com.javaguru.shoppinglist.service.ProductService;
+import com.javaguru.shoppinglist.service.validator.*;
+import com.javaguru.shoppinglist.consoleUI.*;
+
 import java.util.ArrayList;
 import java.util.List;
 
 class ShoppingListApplication {
 
     public static void main(String[] args) {
+        InMemoryDatabase database = new InMemoryDatabase();
         Validation[] validations = new Validation[]{
                 new MinPrice(),
                 new MaxDiscount(),
@@ -14,7 +20,7 @@ class ShoppingListApplication {
         };
 
         ProductValidator productValidator = new ProductValidator(validations);
-        ProductService productService = new DefaultProductService(productValidator);
+        ProductService productService = new DefaultProductService(database, productValidator);
 
         Action exitAction = new ExitAction();
         Action createUserAction = new CreateProductAction(productService);
