--- src/com/javaguru/shoppinglist/Validator/Validation.java
@@ -1,7 +1,0 @@
-package com.javaguru.shoppinglist.Validator;
-
-import com.javaguru.shoppinglist.Product;
-
-public interface Validation {
-    void validate (Product product);
-}
