+++ src/com/javaguru/shoppinglist/consoleUI/Action.java
@@ -1,4 +1,4 @@
-package com.javaguru.shoppinglist;
+package com.javaguru.shoppinglist.consoleUI;
 
 public interface Action {
