+++ src/com/javaguru/shoppinglist/consoleUI/ExitAction.java
@@ -1,4 +1,4 @@
-package com.javaguru.shoppinglist;
+package com.javaguru.shoppinglist.consoleUI;
 
 public class ExitAction implements Action {
