+++ src/com/javaguru/shoppinglist/service/validator/FieldsValidationException.java
@@ -1,4 +1,4 @@
-package com.javaguru.shoppinglist.Validator;
+package com.javaguru.shoppinglist.service.validator;
 
 public class FieldsValidationException extends RuntimeException{
     public FieldsValidationException(String message){
