+++ src/com/javaguru/shoppinglist/service/validator/NameLength.java
@@ -1,6 +1,6 @@
-package com.javaguru.shoppinglist.Validator;
+package com.javaguru.shoppinglist.service.validator;
 
-import com.javaguru.shoppinglist.Product;
+import com.javaguru.shoppinglist.database.Product;
 
 public class NameLength extends AbstractValidator{
     private final int minNameLength = 3;
