+++ src/com/javaguru/shoppinglist/service/validator/Validation.java
@@ -1,0 +1,7 @@
+package com.javaguru.shoppinglist.service.validator;
+
+import com.javaguru.shoppinglist.database.Product;
+
+public interface Validation {
+    void validate (Product product);
+}
